Add a way to hide any dropdown menu opened
Adds an eventListener to window in order to close any dropdown menu that
may be open and prevent an issue where, once one of the dropdowns was
revealed, one dropdown would always remain open with no way to toggle it
off. Additional styling to the nav bar is also introduced.